refactor: 로또 생성을 전략 패턴으로 변경
