+++ src/main/kotlin/lotto/domain/generator/LottoGenerator.kt
@@ -1,0 +1,10 @@
+package lotto.domain.generator
+
+import lotto.domain.Lotto
+
+class LottoGenerator {
+
+    fun generate(lottoCount: Int, lottoGeneratorStrategy: LottoGeneratorStrategy): List<Lotto> {
+        return lottoGeneratorStrategy.generate(lottoCount)
+    }
+}
