+++ src/main/kotlin/lotto/domain/generator/LottoGeneratorStrategy.kt
@@ -1,0 +1,7 @@
+package lotto.domain.generator
+
+import lotto.domain.Lotto
+
+interface LottoGeneratorStrategy {
+    fun generate(lottoCount: Int): List<Lotto>
+}
