+++ src/test/kotlin/lotto/domain/generator/LottoAutoStrategyTest.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.generator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
@@ -6,18 +6,18 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 
-@DisplayName("로또 생성기 테스트")
-class LottoGeneratorTest {
+@DisplayName("로또 자동 생성 테스트")
+class LottoAutoStrategyTest {
 
     @ParameterizedTest
     @ValueSource(ints = [1, 3, 14, 50])
     @DisplayName("구매한 로또 개수만큼 로또를 생성한다")
     fun `sut returns lottos made with lotto count`(lottoCount: Int) {
         // Arrange
-        val sut = LottoGenerator
+        val sut = LottoAutoStrategy()
 
         // Act
-        val lottos = sut.generateAutoLotto(lottoCount)
+        val lottos = sut.generate(lottoCount)
 
         // Assert
         assertThat(lottos).hasSize(lottoCount)
@@ -27,10 +27,10 @@ class LottoGeneratorTest {
     @DisplayName("로또 번호는 1 ~ 45의 번호로 이루어져 있다")
     fun `sut returns lotto number size is 45`() {
         // Arrange
-        val sut = LottoGenerator
+        val sut = LottoAutoStrategy()
 
         // Act
-        val lottos = sut.generateAutoLotto(1)
+        val lottos = sut.generate(1)
 
         // Assert
         val lottoNumbers = lottos[0].lottoNumbers
